feat: do error handling in batch

+++ controller/v1/admin/adminController.js
@@ -1,51 +1,51 @@
 
 import Admin from "../../../models/admin.js";
 import jwt from 'jsonwebtoken';
+import httpError from "../../../utils/httpError.js";
+import bcrypt from 'bcrypt'
 
-const jwt_secret = process.env.JWT 
+const JWT_SECRET = process.env.JWT_SECRET || "5?#562@";
 
 //admin login
 
 
-export const adminLogin =async(req, res, next) =>{
+export const adminLogin = async( req, res, next ) =>{
 
-    try {
+   
+           try {
 
+            const { email,password }= req.body
 
-        const { email, password } =req.body 
-        const admin = await Admin.findOne( { email })
-       const token =jwt.sign(
-        { id: admin._id , role: admin.role } , process.env.JWT 
+            if( !email || !password ) {
 
-       )
-       res.status(200).json({message:"hai"},token)
-    } catch (error) {
-        res.status(500)
+                return next ( new httpError (" email and password Required!") )
+            }
+    
+            const admin = await Admin.findOne( { email } )
+            
+            if( !admin ) {
 
-        console.log(error);
-        
-    }
+                return next ( new httpError ( "Admin not found ", 400 ))
+            }
+            
+            const passwordValid = await bcrypt.compare(password, admin.password);
 
+           
+            if (!passwordValid) {
+                return next( new httpError ( "invalid credentials", 402))
+            }
+          
+            //jwt token 
+            const token = jwt.sign({ id: admin.id, role: admin.role }, JWT_SECRET, { expiresIn: '24h' });
 
+            res.json({message: "Login successfull" , token})
+
+           } catch (error) {
+
+             return next ( new httpError ("Failed to login",500))
+           }
 
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 
@@ -55,16 +55,59 @@ export const adminLogin =async(req, res, next) =>{
 export const addAdmin = async (req, res, next) => {
 
     try {
-      const { first_name, last_name, email, dob, phone, status, password, role,age } = req.body;
-      let profilePicturePath
 
+       //age logic ( take age from dob )
+
+      const calculateAge = (dob) => {
+        const today = new Date();
+        const birthDate = new Date(dob);
+        let age = today.getFullYear() - birthDate.getFullYear();
+        const month = today.getMonth();
+        const day = today.getDate();
+      
+        if (month < birthDate.getMonth() || (month === birthDate.getMonth() && day < birthDate.getDate())) {
+          age--;
+        }
+
+        return age;
+      };
+       
+
+      //body feild
+      const { first_name, last_name, email, dob, phone, status, password, role } = req.body;
+       
+      //validation check
+      if( !first_name || !last_name || !email || !dob || !phone || !status || !password || !role ){
+        return next (new httpError("All credentials are Required!"))
+      }
+
+      //check the admin allready exist
+   
+      const adminExsist = await Admin.findOne({ $or: [{ email }, { phone }] })
+    
+      if (adminExsist) {
+          let errorMessage = adminExsist.email === email ? "An admin with this email already exists" : "An admin with this phone number already exists";
+          return next(new httpError(errorMessage, 407));
+      }
+
+
+      //add admin detailes
+     
+
+      const age = calculateAge(dob)
+
+      const hashedPassword = await bcrypt.hash(password, parseInt(process.env.SALT_VALUE));
+
+      let profilePicturePath
+      
       if(req.file){
          profilePicturePath = req.file.path.slice(8);
       } 
        
         const newAdmin = new Admin(
-            { first_name, last_name, email, dob, phone, status, password, role, profile_pic:profilePicturePath,age} 
+            { first_name, last_name, email, dob, phone, status, password: hashedPassword, role, profile_pic:profilePicturePath, age} 
          );
+
         await newAdmin.save();
 
         res.status(201).json( { message: 'admin created successfully', data: newAdmin } );
@@ -72,14 +115,14 @@ export const addAdmin = async (req, res, next) => {
     }   catch (error) {
         console.log(error)
 
-        res.status(500).json( { message: 'Error creating user', error: error.message } );
+        return next(new httpError("Failed to create admin",500))
     }
   };
 
 
 
 
-//   //find admin
+  //find admin
 
   
 export const listAdmin = async( req, res, next) =>{
@@ -91,7 +134,7 @@ export const listAdmin = async( req, res, next) =>{
       
 
     } catch (error) {
-        next(new Error("Error finding user: " + error.message));
+        return next (new httpError(" Server Error " ,500 ));
     }
 
 }
@@ -101,21 +144,27 @@ export const listAdmin = async( req, res, next) =>{
 
 export const getOneAdmin = async( req, res, next) =>{
 
-    const { id } = req.params;
+    try {
+        const { id } = req.params;
 
     if (!id) {
-        res.status(400).json( { message:` ID error!` } )
+        return next (new httpError("Error finding admin",400))
     }
 
         const admin = await Admin.findOne( {_id: id} );
 
         if (!admin) {
 
-            res.status(404).json( { message:` ADMIN error!` } )
+            return next (new httpError("Admin not present!",402))
         }else {
 
             res.status(200).json( { message:`admin is found` , data : admin} )
         }
+        
+    } catch (error) {
+        
+        return next (new httpError("Internal Server Error",500))
+    }
 
 }
 
@@ -131,6 +180,10 @@ export const updateAdminDetailes = async (req, res, next) =>{
     
     try {
         const { id } = req.params;
+
+        if (!id) {
+            return next (new httpError("Error finding admin",400))
+        }
         const AdminData = {...req.body};
         
         if(req.file && req.file.path){
@@ -147,7 +200,7 @@ export const updateAdminDetailes = async (req, res, next) =>{
 
         if (!updateAdmin) {
 
-            res.status(400).json({ message:`admin not found` })
+           return next (new httpError("Admin not updated!",400))
 
         } else {
 
@@ -155,35 +208,37 @@ export const updateAdminDetailes = async (req, res, next) =>{
         }
 
     }   catch (error) {
-        next( new Error("Error : " + error.message) );
-
-        res.status(500).json({ message:`Internal server error!` })
+        return next (new httpError("Server Error",400))
     }
 }
 
 
 
 
-// //findOneAndDelete
+ //findOneAndDelete
 
 
 
 export const deleteAdmin = async ( req, res, next)=>{
     try {
         const {id} = req.params;
+
+        if (!id) {
+            return next (new httpError("Error finding admin",400))
+        }
         const deleteOneAdmin = await Admin.findOneAndDelete(
             { _id: id }
         )
 
         if ( !deleteOneAdmin ) {
 
-            res.status(400).json({ message:`admin not find`, data:deleteOneAdmin })
+           return next (new httpError("No admin found!",400))
         }
 
             res.status(202).json({message:`admin deleted successfully` ,data:deleteOneAdmin})
 
     }   catch (error) {
-        next(new Error( "error deleting admin :" + error.message ))
+        return next (new httpError("Internal Server Error!",500))
     }
 }
